committing all changes i had done and fields and files added

+++ inventory/src/main/java/com/restaurant/inventory/controller/ItemController.java
@@ -9,16 +9,21 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.restaurant.inventory.model.Categories;
 import com.restaurant.inventory.model.Item;
 import com.restaurant.inventory.model.Unit;
 import com.restaurant.inventory.service.CategoryService;
 import com.restaurant.inventory.service.ItemService;
+import com.restaurant.inventory.service.PurchaseOrderService;
 import com.restaurant.inventory.service.UnitService;
 
 @Controller
 public class ItemController {
+
+    @Autowired
+    private PurchaseOrderService purchaseOrderService;
 
     @Autowired
     private CategoryService categoryService;
@@ -29,7 +34,8 @@ public class ItemController {
     @Autowired
     private ItemService itemService;
 
-    @GetMapping("/item")
+    //method to show the item definition form 
+    @GetMapping("/inventory/item")
     public String showitemform(Model model){
         Item item =new Item();
         model.addAttribute("item", item);
@@ -48,13 +54,14 @@ public class ItemController {
     }
 
     @PostMapping("/saveItem")
-    public String saveItem(@ModelAttribute("item") Item item){
+    public String saveItem(@ModelAttribute("item") Item item,RedirectAttributes redirectAttributes){
         System.out.println(item);
         itemService.defineItem(item);
-        return "redirect:/item";
+        redirectAttributes.addFlashAttribute("successMessage", "Item created successfully.you can link it to supplier in the supplier linkage form");
+        return "redirect:/inventory/item";
     }
 
-    @GetMapping("/itemUpdate/{itemID}")
+    @GetMapping("/inventory/itemUpdate/{itemID}")
     public String showItemUpdateForm(@PathVariable("itemID")int itemID, Model model){
         Item item = itemService.findItemById(itemID);
         model.addAttribute("item", item);
@@ -76,9 +83,23 @@ public class ItemController {
     }
 
     @PostMapping("/deleteItem/{itemID}")
-    public String deleteItem(@PathVariable("itemID")int itemID){
+    public String deleteItem(@PathVariable("itemID")int itemID, RedirectAttributes redirectAttributes){
+        Item item = itemService.findItemById(itemID);
+
+        // Check if item still has stock
+    if (item.getQuantity() > 0) {
+        redirectAttributes.addFlashAttribute("errorMessage", "Cannot delete. Item is still in stock.");
+        return "redirect:/inventory";
+    }
+    // Check if item is involved in a transaction
+    boolean isInTransaction = purchaseOrderService.isItemInTransaction(item);
+    if (isInTransaction) {
+        redirectAttributes.addFlashAttribute("errorMessage", "Cannot delete. Item is involved in a purchaseOrder transaction.");
+        return "redirect:/inventory";
+    }
         itemService.deleteItemById(itemID);
-        return "redirect:/item";
+        redirectAttributes.addFlashAttribute("successMessage", "Item deleted successfully.");
+        return "redirect:/inventory";
     }
 
 }
